Task 10 - Change size_t in int

+++ 10-binary_tree_depth.c
@@ -7,8 +7,8 @@
  */
 size_t binary_tree_depth(const binary_tree_t *tree)
 {
-	size_t left_height;
-	size_t right_height;
+	int left_height;
+	int right_height;
 
 	if (!tree || (!tree->left && !tree->right))
 		return (0);
